+++ src/elements/b_box.rs
@@ -1,6 +1,6 @@
 use leptos::*;
 
 #[component]
-pub fn BBox(children: Children, #[prop(default = "")] class: &'static str) -> impl IntoView {
-    view! { <div class=format!("box {}", class)>{children()}</div> }
+pub fn BBox(children: Children, #[prop(optional, into)] class: TextProp) -> impl IntoView {
+    view! { <div class=format!("box {}", class.get())>{children()}</div> }
 }