+++ website/src/pages/mod.rs
@@ -1,3 +1,4 @@
+mod columns_page;
 mod components_page;
 mod elements_page;
 mod form_page;
@@ -5,6 +6,7 @@ mod guides_page;
 mod home_page;
 mod not_found_page;
 
+pub use columns_page::ColumnsPage;
 pub use components_page::ComponentsPage;
 pub use elements_page::ElementsPage;
 pub use form_page::FormPage;